+++ zerver/models/streams.py
@@ -132,6 +132,11 @@ class Stream(models.Model):
     # stream based on what messages they have cached.
     first_message_id = models.IntegerField(null=True, db_index=True)
 
+    LAST_ACTIVITY_DAYS_BEFORE_FOR_ACTIVE = 180
+
+    # Whether a message has been sent to this stream in the last X days.
+    is_recently_active = models.BooleanField(default=True, db_default=True)
+
     stream_permission_group_settings = {
         "can_remove_subscribers_group": GroupPermissionSetting(
             require_system_group=False,
@@ -184,6 +189,7 @@ class Stream(models.Model):
         "rendered_description",
         "stream_post_policy",
         "can_remove_subscribers_group_id",
+        "is_recently_active",
     ]
 
     def to_dict(self) -> DefaultStreamDict:
@@ -203,6 +209,7 @@ class Stream(models.Model):
             stream_id=self.id,
             stream_post_policy=self.stream_post_policy,
             is_announcement_only=self.stream_post_policy == Stream.STREAM_POST_POLICY_ADMINS,
+            is_recently_active=self.is_recently_active,
         )